scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "frontmatter>=3.0.8",
    "mistune>=3.1.3",
    "pillow>=11.2.1",
]
